tracker/requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
curl_cffi==0.10.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.1.0
python-telegram-bot==22.0
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
